requirements.txt
discord.py>=2.3.2
requests>=2.31.0
beautifulsoup4>=4.12.2
aiohttp>=3.9.1
asyncio>=3.4.3
python-dotenv>=1.0.0
sqlite3
schedule>=1.2.0
lxml>=4.9.3
